feat(editor): Use new sidebar style on Chat hub
