+++ packages/frontend/@n8n/i18n/src/locales/en.json
@@ -456,6 +456,7 @@
 	"chatHub.sidebar.link.newChat": "New chat",
 	"chatHub.sidebar.link.workflowAgents": "Workflow agents",
 	"chatHub.sidebar.link.personalAgents": "Personal agents",
+	"chatHub.sidebar.loadMoreSessions": "Load more",
 	"chatEmbed.infoTip.description": "Add chat to external applications using the n8n chat package.",
 	"chatEmbed.infoTip.link": "More info",
 	"chatEmbed.title": "Embed Chat in your website",
