+++ packages/frontend/editor-ui/src/features/ai/chatHub/ChatWorkflowAgentsView.vue
@@ -14,7 +14,6 @@ import { useMediaQuery } from '@vueuse/core';
 import { MOBILE_MEDIA_QUERY } from '@/features/ai/chatHub/constants';
 import { useRouter } from 'vue-router';
 import ChatLayout from '@/features/ai/chatHub/components/ChatLayout.vue';
-import ChatSidebarOpener from '@/features/ai/chatHub/components/ChatSidebarOpener.vue';
 import SkeletonAgentCard from '@/features/ai/chatHub/components/SkeletonAgentCard.vue';
 import { useI18n } from '@n8n/i18n';
 
@@ -96,7 +95,6 @@ watch(
 				/>
 			</div>
 		</div>
-		<ChatSidebarOpener :class="$style.menuButton" />
 	</ChatLayout>
 </template>
