+++ packages/frontend/editor-ui/src/features/ai/chatHub/components/ChatConversationHeader.vue
@@ -1,9 +1,7 @@
 <script setup lang="ts">
 import { hasPermission } from '@/app/utils/rbac/permissions';
 import type { CredentialsMap } from '@/features/ai/chatHub/chat.types';
-import ChatSidebarOpener from '@/features/ai/chatHub/components/ChatSidebarOpener.vue';
 import ModelSelector from '@/features/ai/chatHub/components/ModelSelector.vue';
-import { useChatHubSidebarState } from '@/features/ai/chatHub/composables/useChatHubSidebarState';
 import { CHAT_VIEW } from '@/features/ai/chatHub/constants';
 import type {
 	ChatHubConversationModel,
@@ -12,14 +10,13 @@ import type {
 	ChatModelDto,
 	ChatSessionId,
 } from '@n8n/api-types';
-import { N8nButton, N8nIconButton } from '@n8n/design-system';
+import { N8nButton } from '@n8n/design-system';
 import { useI18n } from '@n8n/i18n';
 import { computed, useTemplateRef, watch, ref } from 'vue';
 import { useRouter } from 'vue-router';
 import { useChatStore } from '../chat.store';
 
-const { isNewSession, selectedModel, credentials, readyToShowModelSelector } = defineProps<{
-	isNewSession: boolean;
+const { selectedModel, credentials, readyToShowModelSelector } = defineProps<{
 	selectedModel: ChatModelDto | null;
 	credentials: CredentialsMap | null;
 	readyToShowModelSelector: boolean;
@@ -34,7 +31,6 @@ const emit = defineEmits<{
 	openWorkflow: [workflowId: string];
 }>();
 
-const sidebar = useChatHubSidebarState();
 const router = useRouter();
 const modelSelectorRef = useTemplateRef('modelSelectorRef');
 const i18n = useI18n();
@@ -60,8 +56,6 @@ function onModelChange(selection: ChatHubConversationModel) {
 }
 
 function onNewChat() {
-	sidebar.toggleOpen(false);
-
 	void router.push({ name: CHAT_VIEW, force: true });
 }
 
@@ -91,18 +85,6 @@ defineExpose({
 <template>
 	<div :class="$style.component">
 		<div :class="$style.grow">
-			<ChatSidebarOpener :class="$style.menuButton" />
-			<N8nIconButton
-				v-if="!sidebar.isStatic.value"
-				:class="$style.menuButton"
-				type="secondary"
-				icon="square-pen"
-				text
-				icon-size="large"
-				:aria-label="i18n.baseText('chatHub.chat.header.button.newChat')"
-				:disabled="isNewSession"
-				@click="onNewChat"
-			/>
 			<ModelSelector
 				v-if="readyToShowModelSelector"
 				ref="modelSelectorRef"
