+++ packages/frontend/editor-ui/src/features/ai/chatHub/components/ChatSidebarContent.vue
@@ -1,40 +1,31 @@
 <script setup lang="ts">
-import MainSidebarUserArea from '@/app/components/MainSidebarUserArea.vue';
 import { useMessage } from '@/app/composables/useMessage';
 import { useToast } from '@/app/composables/useToast';
-import { MODAL_CONFIRM, VIEWS } from '@/app/constants';
+import { MODAL_CONFIRM } from '@/app/constants';
 import { useChatStore } from '@/features/ai/chatHub/chat.store';
+import { useCredentialsStore } from '@/features/credentials/credentials.store';
 import { groupConversationsByDate } from '@/features/ai/chatHub/chat.utils';
-import ChatSidebarLink from '@/features/ai/chatHub/components/ChatSidebarLink.vue';
-import { useChatHubSidebarState } from '@/features/ai/chatHub/composables/useChatHubSidebarState';
 import {
 	CHAT_VIEW,
 	CHAT_WORKFLOW_AGENTS_VIEW,
 	CHAT_PERSONAL_AGENTS_VIEW,
 } from '@/features/ai/chatHub/constants';
-import { useSettingsStore } from '@/app/stores/settings.store';
-import { N8nIconButton, N8nScrollArea, N8nText } from '@n8n/design-system';
-import Logo from '@n8n/design-system/components/N8nLogo';
-import BetaTag from '@n8n/design-system/components/BetaTag/BetaTag.vue';
+import { type IMenuItem, N8nMenuItem, N8nScrollArea, N8nText } from '@n8n/design-system';
 import { computed, onMounted, ref } from 'vue';
 import { useRoute, useRouter } from 'vue-router';
-import { useIntersectionObserver } from '@vueuse/core';
 import ChatSessionMenuItem from './ChatSessionMenuItem.vue';
 import SkeletonMenuItem from './SkeletonMenuItem.vue';
 import { useTelemetry } from '@/app/composables/useTelemetry';
 import { type ChatHubSessionDto } from '@n8n/api-types';
-import { useCredentialsStore } from '@/features/credentials/credentials.store';
 import { useI18n } from '@n8n/i18n';
 
-defineProps<{ isMobileDevice: boolean }>();
+defineProps<{ isMobileDevice: boolean; isCollapsed: boolean }>();
 
 const route = useRoute();
 const router = useRouter();
 const chatStore = useChatStore();
 const toast = useToast();
 const message = useMessage();
-const sidebar = useChatHubSidebarState();
-const settingsStore = useSettingsStore();
 const credentialsStore = useCredentialsStore();
 const telemetry = useTelemetry();
 const readyToShowSessions = computed(
@@ -43,7 +34,6 @@ const readyToShowSessions = computed(
 const i18n = useI18n();
 
 const renamingSessionId = ref<string>();
-const loadMoreTrigger = ref<HTMLElement | null>(null);
 
 const currentSessionId = computed(() =>
 	typeof route.params.id === 'string' ? route.params.id : undefined,
@@ -62,6 +52,40 @@ const groupedConversations = computed(() =>
 		}, []),
 	),
 );
+
+const newChat = computed<IMenuItem>(() => ({
+	id: 'new-chat',
+	label: i18n.baseText('chatHub.sidebar.link.newChat'),
+	icon: 'plus',
+	route: {
+		to: {
+			name: CHAT_VIEW,
+			force: true, // to focus input again when the user is already in CHAT_VIEW
+		},
+	},
+}));
+
+const personalAgents = computed<IMenuItem>(() => ({
+	id: 'personal-agents',
+	label: i18n.baseText('chatHub.sidebar.link.personalAgents'),
+	icon: 'message-square',
+	route: {
+		to: {
+			name: CHAT_PERSONAL_AGENTS_VIEW,
+		},
+	},
+}));
+
+const workflowAgents = computed<IMenuItem>(() => ({
+	id: 'workflow-agents',
+	label: i18n.baseText('chatHub.sidebar.link.workflowAgents'),
+	icon: 'robot',
+	route: {
+		to: {
+			name: CHAT_WORKFLOW_AGENTS_VIEW,
+		},
+	},
+}));
 
 function handleStartRename(sessionId: string) {
 	renamingSessionId.value = sessionId;
@@ -108,18 +132,7 @@ async function handleDeleteSession(sessionId: string) {
 
 function handleNewChatClick() {
 	telemetry.track('User clicked new chat button', {});
-	sidebar.toggleOpen(false);
 }
-
-useIntersectionObserver(
-	loadMoreTrigger,
-	([{ isIntersecting }]) => {
-		if (isIntersecting) {
-			void chatStore.fetchMoreSessions({ minLoadingTime: 250 });
-		}
-	},
-	{ threshold: 0.1 },
-);
 
 onMounted(() => {
 	if (!chatStore.sessionsReady) {
@@ -130,73 +143,56 @@ onMounted(() => {
 
 <template>
 	<div :class="[$style.component, { [$style.isMobileDevice]: isMobileDevice }]">
-		<div :class="$style.header">
-			<RouterLink :to="{ name: VIEWS.HOMEPAGE }">
-				<div :class="$style.logoContainer">
-					<Logo
-						:class="$style.logo"
-						size="small"
-						:collapsed="false"
-						:release-channel="settingsStore.settings.releaseChannel"
-					/>
-					<BetaTag />
-				</div>
-			</RouterLink>
-			<N8nIconButton
-				v-if="sidebar.canBeStatic.value"
-				:title="i18n.baseText('chatHub.sidebar.button.toggle')"
-				icon="panel-left"
-				type="tertiary"
-				text
-				size="small"
-				icon-size="large"
-				@click="sidebar.toggleStatic()"
-			/>
-		</div>
-		<div :class="$style.links">
-			<ChatSidebarLink
-				:to="{
-					name: CHAT_VIEW,
-					force: true, // to focus input again when the user is already in CHAT_VIEW
-				}"
-				:label="i18n.baseText('chatHub.sidebar.link.newChat')"
-				icon="square-pen"
+		<div
+			:class="{
+				[$style.links]: true,
+				[$style.collapsed]: isCollapsed,
+			}"
+		>
+			<N8nMenuItem
+				:item="newChat"
+				:compact="isCollapsed"
 				:active="route.name === CHAT_VIEW"
 				@click="handleNewChatClick"
 			/>
-			<ChatSidebarLink
-				:to="{ name: CHAT_PERSONAL_AGENTS_VIEW }"
-				:label="i18n.baseText('chatHub.sidebar.link.personalAgents')"
-				icon="message-square"
+			<N8nMenuItem
+				:item="personalAgents"
+				:compact="isCollapsed"
 				:active="route.name === CHAT_PERSONAL_AGENTS_VIEW"
 				@click="sidebar.toggleOpen(false)"
 			/>
-			<ChatSidebarLink
-				:to="{ name: CHAT_WORKFLOW_AGENTS_VIEW }"
-				:label="i18n.baseText('chatHub.sidebar.link.workflowAgents')"
-				icon="robot"
+			<N8nMenuItem
+				:item="workflowAgents"
+				:compact="isCollapsed"
 				:active="route.name === CHAT_WORKFLOW_AGENTS_VIEW"
 				@click="sidebar.toggleOpen(false)"
 			/>
 		</div>
 		<N8nScrollArea as-child type="scroll">
-			<div :class="$style.historySections">
+			<div :class="[$style.historySections, { [$style.collapsed]: isCollapsed }]">
 				<div v-if="!readyToShowSessions" :class="$style.group">
 					<SkeletonMenuItem v-for="i in 10" :key="`loading-${i}`" />
 				</div>
 				<div
-					v-else
 					v-for="(group, index) in groupedConversations"
+					v-else
 					:key="group.group"
 					:class="$style.group"
 				>
-					<N8nText :class="$style.groupHeader" size="small" bold color="text-light">
+					<N8nText
+						v-if="!isCollapsed"
+						:class="$style.groupHeader"
+						size="small"
+						bold
+						color="text-light"
+					>
 						{{ group.group }}
 					</N8nText>
 					<ChatSessionMenuItem
 						v-for="session in group.sessions"
 						:key="session.id"
 						:session="session"
+						:compact="isCollapsed"
 						:active="currentSessionId === session.id"
 						:is-renaming="renamingSessionId === session.id"
 						@start-rename="handleStartRename"
@@ -204,15 +200,27 @@ onMounted(() => {
 						@confirm-rename="handleConfirmRename"
 						@delete="handleDeleteSession"
 					/>
+
+					<N8nMenuItem
+						v-if="
+							index === groupedConversations.length - 1 &&
+							chatStore.sessions.hasMore &&
+							!chatStore.sessionsLoading
+						"
+						:item="{
+							id: 'load-more-sessions',
+							label: $t('chatHub.sidebar.loadMoreSessions'),
+							icon: 'circle-ellipsis',
+						}"
+						:compact="isCollapsed"
+						@click="() => chatStore.fetchMoreSessions({ minLoadingTime: 250 })"
+					/>
 					<template v-if="index === groupedConversations.length - 1 && chatStore.sessionsLoading">
 						<SkeletonMenuItem v-for="i in 10" :key="i" />
 					</template>
 				</div>
-
-				<div ref="loadMoreTrigger" :class="$style.loadMoreTrigger"></div>
 			</div>
 		</N8nScrollArea>
-		<MainSidebarUserArea :fully-expanded="true" :is-collapsed="false" />
 	</div>
 </template>
 
@@ -253,18 +261,26 @@ onMounted(() => {
 .links {
 	display: flex;
 	flex-direction: column;
-	padding: 0 var(--spacing--xs) var(--spacing--sm) var(--spacing--xs);
+	padding: 0 var(--spacing--3xs) var(--spacing--2xs);
 	gap: 1px;
+
+	&.collapsed {
+		border-bottom: var(--border);
+	}
 }
 
 .historySections {
 	display: flex;
 	flex-direction: column;
-	padding: 0 var(--spacing--xs) var(--spacing--sm) var(--spacing--xs);
+	padding: var(--spacing--2xs) var(--spacing--3xs) var(--spacing--2xs);
 	gap: var(--spacing--xs);
 
 	.isMobileDevice & {
 		gap: var(--spacing--sm);
+	}
+
+	&.collapsed {
+		gap: 0;
 	}
 }
 
@@ -276,11 +292,6 @@ onMounted(() => {
 
 .groupHeader {
 	padding: 0 var(--spacing--4xs) var(--spacing--3xs) var(--spacing--4xs);
-}
-
-.loadMoreTrigger {
-	height: 1px;
-	width: 100%;
 }
 
 .loading,
