+++ packages/frontend/editor-ui/src/features/collaboration/projects/components/ProjectNavigation.vue
@@ -17,6 +17,7 @@ import { hasPermission } from '@/app/utils/rbac/permissions';
 type Props = {
 	collapsed: boolean;
 	planName?: string;
+	hideProjects?: boolean;
 };
 
 const props = defineProps<Props>();
@@ -146,34 +147,38 @@ onBeforeUnmount(() => {
 				data-test-id="project-chat-menu-item"
 			/>
 		</div>
-		<N8nText
-			v-if="
-				!props.collapsed && projectsStore.isTeamProjectFeatureEnabled && displayProjects.length > 0
-			"
-			:class="[$style.projectsLabel]"
-			size="small"
-			bold
-			role="heading"
-			color="text-light"
-		>
-			{{ locale.baseText('projects.menu.title') }}
-		</N8nText>
-		<div
-			v-if="projectsStore.isTeamProjectFeatureEnabled || isFoldersFeatureEnabled"
-			:class="$style.projectItems"
-		>
-			<N8nMenuItem
-				v-for="project in displayProjects"
-				:key="project.id"
-				:class="{
-					[$style.collapsed]: props.collapsed,
-				}"
-				:item="getProjectMenuItem(project)"
-				:compact="props.collapsed"
-				:active="activeTabId === project.id"
-				data-test-id="project-menu-item"
-			/>
-		</div>
+		<template v-if="!props.hideProjects">
+			<N8nText
+				v-if="
+					!props.collapsed &&
+					projectsStore.isTeamProjectFeatureEnabled &&
+					displayProjects.length > 0
+				"
+				:class="[$style.projectsLabel]"
+				size="small"
+				bold
+				role="heading"
+				color="text-light"
+			>
+				{{ locale.baseText('projects.menu.title') }}
+			</N8nText>
+			<div
+				v-if="projectsStore.isTeamProjectFeatureEnabled || isFoldersFeatureEnabled"
+				:class="$style.projectItems"
+			>
+				<N8nMenuItem
+					v-for="project in displayProjects"
+					:key="project.id"
+					:class="{
+						[$style.collapsed]: props.collapsed,
+					}"
+					:item="getProjectMenuItem(project)"
+					:compact="props.collapsed"
+					:active="activeTabId === project.id"
+					data-test-id="project-menu-item"
+				/>
+			</div>
+		</template>
 	</div>
 </template>
